+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * feat(Checkbox/Switch): add support for `trueValue` / `falseValue`
 * feat(FileUpload): add `fileImage` prop
 * feat(ChatTool): new component
+* feat(Table): implement row pinning
 * 
 * fix(Error): support `status` and `statusText` properties
 * fix(ContentSurround): handle RTL mode